Modification de l'API et de l'application JFX pour permettre aux utilisateurs de créer et modifier des évaluations
L'utilisateur peut désormais, ajouter une nouvelle évaluation ou modifier une existante (via l'interface JFX).

+++ gnoteFX1.2/src/main/java/org/openjfx/sio2E4/controller/EtudiantsController.java
@@ -52,7 +52,7 @@ public class EtudiantsController {
                                     "ETUDIANT".equalsIgnoreCase(u.getRole().getLibelle()))
                             .map(u -> {
                                 Etudiant e = new Etudiant();
-                                e.setId((long) u.getId());
+                                e.setId(u.getId());
                                 e.setNom(u.getNom());
                                 e.setPrenom(u.getPrenom());
                                 e.setEmail(u.getEmail());
